Add rsrec(), rsrecnext(0, rsrecprev(), rsnext(), rsprev().

+++ mux/src/htab.h
@@ -47,7 +47,7 @@ struct bque
     reg_ref *scr[MAX_GLOBAL_REGS];  // temp vars
 #if defined(STUB_SLAVE)
     CResultsSet *pResultsSet;       // Results Set
-    size_t  iRow;                   // Current Row
+    int     iRow;                   // Current Row
 #endif // STUB_SLAVE
     bool    IsTimed;                // Is there a waittime time on this entry?
 };
